+++ adyenocc/src/com/adyen/commerce/controllers/PlaceOrderController.java
@@ -3,10 +3,15 @@ package com.adyen.commerce.controllers;
 import com.adyen.commerce.constants.AdyenoccConstants;
 import com.adyen.commerce.controllerbase.PlaceOrderControllerBase;
 import com.adyen.commerce.facades.AdyenCheckoutApiFacade;
+import com.adyen.commerce.request.PlaceOrderRequest;
 import com.adyen.commerce.response.PlaceOrderResponse;
 import com.adyen.model.checkout.PaymentDetailsRequest;
 import com.fasterxml.jackson.core.JsonProcessingException;
+import de.hybris.platform.acceleratorfacades.flow.CheckoutFlowFacade;
+import de.hybris.platform.acceleratorservices.urlresolver.SiteBaseUrlResolutionService;
+import de.hybris.platform.commercefacades.order.CartFacade;
 import de.hybris.platform.commerceservices.request.mapping.annotation.ApiVersion;
+import de.hybris.platform.site.BaseSiteService;
 import de.hybris.platform.webservicescommons.swagger.ApiBaseSiteIdUserIdAndCartIdParam;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -19,6 +24,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
+
 @Controller
 @RequestMapping(value = AdyenoccConstants.ADYEN_USER_CART_PREFIX)
 @ApiVersion("v2")
@@ -28,9 +36,31 @@ public class PlaceOrderController extends PlaceOrderControllerBase {
     @Autowired
     private AdyenCheckoutApiFacade adyenCheckoutApiFacade;
 
+    @Autowired
+    private CheckoutFlowFacade checkoutFlowFacade;
 
-    @Secured({ "ROLE_CUSTOMERGROUP", "ROLE_CLIENT", "ROLE_CUSTOMERMANAGERGROUP", "ROLE_TRUSTED_CLIENT" })
-    @PostMapping(value = "/additional-details", produces= MediaType.APPLICATION_JSON_VALUE)
+    @Autowired
+    private CartFacade cartFacade;
+
+    @Resource(name = "siteBaseUrlResolutionService")
+    private SiteBaseUrlResolutionService siteBaseUrlResolutionService;
+
+    @Resource(name = "baseSiteService")
+    private BaseSiteService baseSiteService;
+
+    @Secured({"ROLE_CUSTOMERGROUP", "ROLE_CLIENT", "ROLE_CUSTOMERMANAGERGROUP", "ROLE_TRUSTED_CLIENT"})
+    @PostMapping(value = "/place-order", produces = MediaType.APPLICATION_JSON_VALUE)
+    @Operation(operationId = "placeOrder", summary = "Handle place order request", description =
+            "Places order based on request data")
+    @ApiBaseSiteIdUserIdAndCartIdParam
+    public ResponseEntity<PlaceOrderResponse> onPlaceOrder(@RequestBody PlaceOrderRequest placeOrderRequest, HttpServletRequest request) throws Exception {
+        PlaceOrderResponse placeOrderResponse = super.placeOrder(placeOrderRequest, request);
+
+        return ResponseEntity.ok(placeOrderResponse);
+    }
+
+    @Secured({"ROLE_CUSTOMERGROUP", "ROLE_CLIENT", "ROLE_CUSTOMERMANAGERGROUP", "ROLE_TRUSTED_CLIENT"})
+    @PostMapping(value = "/additional-details", produces = MediaType.APPLICATION_JSON_VALUE)
     @Operation(operationId = "additionalDetails", summary = "Handle additional details action", description =
             "Places pending order based on additional details request")
     @ApiBaseSiteIdUserIdAndCartIdParam
@@ -44,5 +74,25 @@ public class PlaceOrderController extends PlaceOrderControllerBase {
     @Override
     public AdyenCheckoutApiFacade getAdyenCheckoutApiFacade() {
         return adyenCheckoutApiFacade;
+    }
+
+    @Override
+    public CheckoutFlowFacade getCheckoutFlowFacade() {
+        return checkoutFlowFacade;
+    }
+
+    @Override
+    public CartFacade getCartFacade() {
+        return cartFacade;
+    }
+
+    @Override
+    public BaseSiteService getBaseSiteService() {
+        return baseSiteService;
+    }
+
+    @Override
+    public SiteBaseUrlResolutionService getSiteBaseUrlResolutionService() {
+        return siteBaseUrlResolutionService;
     }
 }
