+++ adyenwebcommons/src/com/adyen/commerce/util/FieldValidationUtil.java
